requirements.txt
google-api-core==2.11.0
google-auth==2.18.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
numpy==1.24.3
pandas==2.0.2
streamlit==1.23.1
